Reformat searching, sorting, & heaps and part of linked lists
