+++ Linked Lists/SublistReversal/SublistReversal.java
@@ -1,104 +1,93 @@
-/*
-  Input:
-    start = 2, end = 4
-    1 -> 2 -> 3 -> 4 -> 5 -> x
+class Solution {
+  public ListNode reverseBetween(ListNode head, int start, int end) {
+    if (start == end) {
+      return head;
+    }
 
-    positions:
-    1 -> 2 -> 3 -> 4 -> 5 -> x
-    1    2    3    4    5
-       start      end
+    // Pad a dummy head to the start of the list
+    ListNode dummyHead = new ListNode(-1);
+    dummyHead.next = head;
 
-  Output:
-    1 -> 4 -> 3 -> 2 -> 5 -> x
-*/
-public ListNode reverseBetween(ListNode head, int start, int end) {
-  if (start == end) {
-    return head;
-  }
+    /*
+      Create a working pointer to advance to right before the sublist head:
 
-  // Pad a dummy head to the start of the list
-  ListNode dummyHead = new ListNode(-1);
-  dummyHead.next = head;
-
-  /*
-    Create a working pointer to advance to right before the sublist head:
-
-    pos = 1
-    dummyHead -> 1 -> 2 -> 3 -> 4 -> 5 -> x
-        ^
-    
-    pos = 2
-    dummyHead -> 1 -> 2 -> 3 -> 4 -> 5 -> x
-                 ^
-    break, pos == start
-  */
-  ListNode nodeBeforeReversedSublist = dummyHead;
-  int pos = 1;
-  while (pos < start) {
-    nodeBeforeReversedSublist = nodeBeforeReversedSublist.next;
-    pos++;
-  }
-
-  /*
-    Reverse the actual sublist:
-
-    start:
+      pos = 1
       dummyHead -> 1 -> 2 -> 3 -> 4 -> 5 -> x
-                  nbrs swp
-    
-    1st iteration (start = 2):
-      1.) Cut out of sublist
+          ^
+      
+      pos = 2
+      dummyHead -> 1 -> 2 -> 3 -> 4 -> 5 -> x
+                  ^
+      break, pos == start
+    */
+    ListNode nodeBeforeReversedSublist = dummyHead;
+    int pos = 1;
+    while (pos < start) {
+      nodeBeforeReversedSublist = nodeBeforeReversedSublist.next;
+      pos++;
+    }
+
+    /*
+      Reverse the actual sublist:
+
+      start:
         dummyHead -> 1 -> 2 -> 3 -> 4 -> 5 -> x
-                    nbrs swp nctsf
-        
-        dummyHead -> 1 -> 2 -> 4 -> 5 -> x
-                    nbrs swp   ^
-                            3--|
-                          nctsf
-
-      2.) Wire into sublist head
-        dummyHead -> 1 -> 2 -> 4 -> 5 -> x
                     nbrs swp
-                          ^ 
-                       3--|
-                     nctsf
-
-        dummyHead -> 1 -> 3 -> 2 -> 4 -> 5 -> x
-                    nbrs      swp
       
-    2nd iteration (start = 3):
-      1.) Cut out of sublist
-        dummyHead -> 1 -> 3 -> 2 -> 4 -> 5 -> x
-                   nbrs       swp nctsf
+      1st iteration (start = 2):
+        1.) Cut out of sublist
+          dummyHead -> 1 -> 2 -> 3 -> 4 -> 5 -> x
+                      nbrs swp nctsf
+          
+          dummyHead -> 1 -> 2 -> 4 -> 5 -> x
+                      nbrs swp   ^
+                              3--|
+                            nctsf
+
+        2.) Wire into sublist head
+          dummyHead -> 1 -> 2 -> 4 -> 5 -> x
+                      nbrs swp
+                            ^ 
+                        3--|
+                      nctsf
+
+          dummyHead -> 1 -> 3 -> 2 -> 4 -> 5 -> x
+                      nbrs      swp
         
-        dummyHead -> 1 -> 3 -> 2 -> 5 -> x
-                   nbrs       swp   ^
-                                 4--|
-                               nctsf
+      2nd iteration (start = 3):
+        1.) Cut out of sublist
+          dummyHead -> 1 -> 3 -> 2 -> 4 -> 5 -> x
+                    nbrs       swp nctsf
+          
+          dummyHead -> 1 -> 3 -> 2 -> 5 -> x
+                    nbrs       swp   ^
+                                  4--|
+                                nctsf
+        
+        2.) Wire into sublist head
+          dummyHead -> 1 -> 3 -> 2 -> 5 -> x
+                    nbrs   ^   swp
+                        4--|
+                      nctsf
+
+          dummyHead -> 1 -> 4 -> 3 -> 2 -> 5 -> x
+                    nbrs            swp
       
-      2.) Wire into sublist head
-        dummyHead -> 1 -> 3 -> 2 -> 5 -> x
-                   nbrs   ^   swp
-                       4--|
-                     nctsf
+      3rd Iteration (start = 4), break while loop
+    */
+    ListNode sublistWorkingPtr = nodeBeforeReversedSublist.next;
+    while (start < end) {
+      // 1.) Cut out of sublist
+      ListNode nodeComingToSublistFront = sublistWorkingPtr.next;
+      sublistWorkingPtr.next = nodeComingToSublistFront.next;
 
-        dummyHead -> 1 -> 4 -> 3 -> 2 -> 5 -> x
-                   nbrs            swp
-    
-    3rd Iteration (start = 4), break while loop
-  */
-  ListNode sublistWorkingPtr = nodeBeforeReversedSublist.next;
-  while (start < end) {
-    // 1.) Cut out of sublist
-    ListNode nodeComingToSublistFront = sublistWorkingPtr.next;
-    sublistWorkingPtr.next = nodeComingToSublistFront.next;
+      // 2.) Wire into sublist head
+      nodeComingToSublistFront.next = nodeBeforeReversedSublist.next;
+      nodeBeforeReversedSublist.next = nodeComingToSublistFront;
 
-    // 2.) Wire into sublist head
-    nodeComingToSublistFront.next = nodeBeforeReversedSublist.next;
-    nodeBeforeReversedSublist.next = nodeComingToSublistFront;
+      start++;
+    }
 
-    start++;
+    return dummyHead.next;
   }
-
-  return dummyHead.next;
 }
