+++ Sorting, Searching, & Heaps/MinimumInCyclicSortedArray/MinimumInCyclicSortedArray.java
@@ -1,16 +1,18 @@
-public int findMin(int[] nums) {
-  int left = 0;
-  int right = nums.length - 1;
+class Solution {
+  public int findMin(int[] nums) {
+    int left = 0;
+    int right = nums.length - 1;
 
-  while (left < right) {
-    int mid = left + ((right - left) / 2);
+    while (left < right) {
+      int mid = left + ((right - left) / 2);
 
-    if (nums[mid] > nums[right]) {
-      left = mid + 1;
-    } else {
-      right = mid;
+      if (nums[mid] > nums[right]) {
+        left = mid + 1;
+      } else {
+        right = mid;
+      }
     }
-  }
 
-  return nums[left];
+    return nums[left];
+  }
 }
