+++ Sorting, Searching, & Heaps/RawBubbleSort/RawBubbleSort.java
@@ -1,16 +1,5 @@
-/*
-  Bubble Sort
-*/
-import java.util.*;
-
-public class RawBubbleSort {
-  public static void main(String args[]) {
-    int[] numbers = new int[]{ 1, -40, 4, 5, 8, 3, 4, 1, 4, 5, 100, -12 };
-    bubbleSort(numbers);
-    System.out.println(Arrays.toString(numbers));
-  }
-
-  private static void bubbleSort(int[] arr) {
+class RawBubbleSort {
+  public void bubbleSort(int[] arr) {
     int lastIndex = arr.length - 1;
 
     for (int i = lastIndex; i >= 1; i--) {
@@ -22,7 +11,7 @@ public class RawBubbleSort {
     }
   }
 
-  private static void swap(int[] arr, int first, int second) {
+  private void swap(int[] arr, int first, int second) {
     int temp = arr[first];
     arr[first] = arr[second];
     arr[second] = temp;
